Module Gettext

Modules to use in libraries and programs.

author Sylvain Le Gall

This module defines all the function required to use gettext. The primary design is to use applicative function. The "side effect" of such a choice is that you must defines, before using any function, all the text domains, codeset et al. When building a library, you should give access to Library.init (by defining a gettext_init = YouLibrary.init). This is required to enable string translation in the library and programs that uses the library. The only function missing here is the realize function. This function is defined in a real implementation library :

GettextDummy

GettextCamomile

GettextStub

Exception

val string_of_exception : exn -> string

Return the string representation of a ocaml-gettext exception.

High level interfaces

val init : GettextTypes.dependencies

Value of the dependencies for the initialization of the library Gettext (for translating exception and help message).

module Library (_ : GettextTypes.INIT_TYPE) : sig ... end

Module to handle typical library requirement

module Program (_ : GettextTypes.INIT_TYPE) (_ : GettextTypes.REALIZE_TYPE) : sig ... end

Module to handle typical program requirement