Module type GettextTranslate.TRANSLATE_TYPE

type u

val create : GettextTypes.t -> GettextTypes.filename -> (string -> string) -> u

create t filename recode : Create a translation table using filename as the mo file and recode as the encoding converter.

val translate : u -> bool -> string -> (string * int) option -> string

translate str (plural_form,number) tbl : translate the string str using tbl. It is possible that the operation modify tbl, so it is returned also. It is also possible to get the plural form of the translated string using plural_form and number.