Module GettextTypes

Types and exception of ocaml-gettext.

author Sylvain Le Gall

Core types of ocaml-gettext library

type range = Stdlib.Int32.t * Stdlib.Int32.t

type textdomain = string

type locale = string

type dir = string

type filename = string

type codeset = string

Exceptions

exception CompileProblemReadingFile of filename * string

Filename wich generates the error message str

exception CompileExtractionFailed of filename * string * int

While extracting filename the command str returns exit code i.

exception CompileExtractionInterrupted of filename * string * int

While extracting filename the command receive signal i.

exception DomainFileDoesntExist of filename list

Cannot the filename corresponding to a textdomain among the specified files.

exception FormatInconsistent of string * string

The two strings returned doesn't have the same meaning regarding Printf syntax.

exception GettextUninitialized

A part of the code try to translate a string, but ocaml-gettext is not initialized.

exception MoInvalidOptions of Stdlib.Lexing.lexbuf * string

There is an invalid field in the content information of a MO file.

exception MoInvalidPlurals of Stdlib.Lexing.lexbuf * string

The plural-form field is not correct.

exception MoInvalidContentType of Stdlib.Lexing.lexbuf * string

The content-type field is not correct.

exception MoInvalidTranslationSingular of string * int

A plural translation of a singular string has occured.

exception MoInvalidTranslationPlural of string list * int

An out-of-bound plural translation has occured.

exception MoJunk of string * string list

There is more plural translation than the number of plural forms.

exception MoEmptyEntry

exception MoInvalidFile

A MO corrupted file has been read.

exception MoInvalidHeaderNegativeStrings

The MO file specified a negative number of strings.

exception MoInvalidHeaderTableStringOutOfBound of range * range

Offset of the string table is out of bound.

exception MoInvalidHeaderTableTranslationOutOfBound of range * range

Offset of the translation table is out of bound.

exception MoInvalidHeaderTableTranslationStringOverlap of range * range

String and translation table overlap.

exception MoInvalidStringOutOfBound of int * int

The offset and length of a string entry leads to an access beyond the end of the MO file.

exception MoInvalidTranslationOutOfBound of int * int

The offset and length of a translation entry leads to an access beyond the end of the MO file.

exception MoCannotOpenFile of string

An error occured when trying to open a MO file.

exception PoInvalidFile of string * Stdlib.Lexing.lexbuf * Stdlib.in_channel

A PO file cannot be parsed.

exception PoFileInvalidIndex of string * int

When parsing a PO file, found an out of order table indices in a plural form.

exception PoFileDoesntExist of string

The PO file doesn't exist.

exception PoInconsistentMerge of string * string

Cannot merge two PO files.

exception TranslateStringNotFound of string

A string to translate cannot be found.

exception LocalePosixUnparseable of string

Cannot parse the POSIX representation of the locale.

Modules signatures

type dependencies = (textdomain * codeset option * dir option) list

module type INIT_TYPE = sig ... end