Module GettextCompat

Gettext compatibility with the native gettext API

author

Sylvain Le Gall

val textdomain : GettextTypes.textdomain -> GettextTypes.t -> GettextTypes.t

textdomain domain t Set the current text domain.

val get_textdomain : GettextTypes.t -> GettextTypes.textdomain

get_textdomain t Returns the current text domain.

val bindtextdomain : GettextTypes.textdomain -> GettextTypes.dir -> GettextTypes.t -> GettextTypes.t

bindtextdomain textdomain dir Set the default base directory for the specified domain.

val bind_textdomain_codeset : GettextTypes.textdomain -> GettextTypes.codeset -> GettextTypes.t -> GettextTypes.t

bind_textdomain_codeset textdomain codeset Set the codeset to use for the specified domain. codeset must be a valid codeset for the underlying character encoder/decoder (iconv, camomile, extlib...)

val gettext : GettextTypes.t' -> string -> string

gettext t' str Translate the string str.

val fgettext : GettextTypes.t' -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fgettext t' str gettext returning format.

val dgettext : GettextTypes.t' -> GettextTypes.textdomain -> string -> string

dgettext t' textdomain str Translate the string str for the specified domain.

val fdgettext : GettextTypes.t' -> GettextTypes.textdomain -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fdgettext t' textdomain str dgettext returning format.

val dcgettext : GettextTypes.t' -> GettextTypes.textdomain -> string -> GettextCategory.category -> string

dcgettext t' textdomain str category Translate the string str for the specified domain and category.

val fdcgettext : GettextTypes.t' -> GettextTypes.textdomain -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> GettextCategory.category -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fdcgettext t' textdomain str category dcgettext returning format.

val ngettext : GettextTypes.t' -> string -> string -> int -> string

ngettext t' str str_plural n Translate the string str using a plural form. str_plural is the default english plural. n is the relevant number for plural (i.e. the number of objects deals with the string).

val fngettext : GettextTypes.t' -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> int -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fngettext t' str str_plural n ngettext returning format.

val dngettext : GettextTypes.t' -> GettextTypes.textdomain -> string -> string -> int -> string

dngettext t' textdomain str str_plural n Translate the string str using a plural form for the specified domain.

val fdngettext : GettextTypes.t' -> GettextTypes.textdomain -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> int -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fdngettext t' textdomain str str_plural n dngettext returning format.

val dcngettext : GettextTypes.t' -> GettextTypes.textdomain -> string -> string -> int -> GettextCategory.category -> string

dcngettext t' textdomain str str_plural n category Translate the string str using a plural form for the specified domain and category.

val fdcngettext : GettextTypes.t' -> GettextTypes.textdomain -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> int -> GettextCategory.category -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

fdcngettext t' textdomain str str_plural n category dcngettext returning format.