Module GettextFormat

Check string equivalence regarding printf use.

author Sylvain Le Gall

val check_format : GettextTypes.failsafe -> GettextTypes.translation -> GettextTypes.translation

check_format failsafe translation Returns a translation structure if all the string contained in the translation are equivalent of str_id, regarding printf format. If not, replace each string which conflict by str_id, in the result.