Module GettextLocale

Implements different operation over locale/category.

author Sylvain Le Gall

module type LOCALE_TYPE = sig ... end

val posix_getenv : GettextTypes.t -> GettextCategory.category -> GettextTypes.locale

Return the best value of environnement variable, that can be found according to the priority defined in gettext. The choice take into account t and category, but may ignore it, if a variable with a best priority is set. This function can be used to get a value for a LOCALE_TYPE implementation. Raise Not_found if nothing appropriate.

module Posix : LOCALE_TYPE